requirements.txt
streamlit>=1.32.0
transformers==4.40.2
torch>=2.0.0
sentencepiece>=0.1.99
tf-keras
Pillow>=10.0.0
pytesseract>=0.3.10
sacremoses